CTCP-5329: Moved common navigation methods to trait.
