+++ app/navigation/LocationOfGoodsNavigator.scala
@@ -19,13 +19,10 @@ package navigation
 import com.google.inject.Singleton
 import config.Constants.QualifierOfTheIdentification._
 import models._
-import navigation.ContainerNavigator.containerIndicatorPageNavigation
-import navigation.LoadingNavigator._
-import navigation.LocationOfGoodsNavigator.limitDatePageNavigator
 import pages._
 import pages.locationOfGoods._
 import pages.locationOfGoods.contact.{NamePage, PhoneNumberPage}
-import pages.transport.{CheckInformationPage, ContainerIndicatorPage, LimitDatePage}
+import pages.transport.{CheckInformationPage, LimitDatePage}
 import play.api.mvc.Call
 
 import javax.inject.Inject
@@ -46,7 +43,7 @@ class LocationOfGoodsNavigator @Inject() () extends Navigator {
     case NamePage                                                                                 => ua => PhoneNumberPage.route(ua, departureId, mode)
     case CustomsOfficeIdentifierPage                                                              => ua => placeOfLoadingExistsRedirect(ua, departureId, mode)
     case PhoneNumberPage                                                                          => ua => phoneNumberPageNavigation(ua, departureId, mode)
-    case LimitDatePage                                                                            => ua => limitDatePageNavigator(departureId, mode, ua)
+    case LimitDatePage                                                                            => ua => limitDatePageNavigation(departureId, mode, ua)
   }
 
   override def checkRoutes(departureId: String, mode: Mode): PartialFunction[Page, UserAnswers => Option[Call]] = {
@@ -63,13 +60,13 @@ class LocationOfGoodsNavigator @Inject() () extends Navigator {
     case CountryPage                                                                              => ua => AddressPage.route(ua, departureId, mode)
   }
 
-  def namePageNavigation(ua: UserAnswers, departureId: String, mode: Mode): Option[Call] =
+  private def namePageNavigation(ua: UserAnswers, departureId: String, mode: Mode): Option[Call] =
     (ua.get(PhoneNumberPage).isEmpty, ua.get(AddContactYesNoPage).contains(true)) match {
       case (true, true) => Some(controllers.locationOfGoods.contact.routes.PhoneNumberController.onPageLoad(departureId, mode))
       case _            => Some(controllers.routes.CheckYourAnswersController.onPageLoad(departureId))
     }
 
-  def routeIdentificationPageNavigation(userAnswers: UserAnswers, departureId: String, mode: Mode): Option[Call] =
+  private def routeIdentificationPageNavigation(userAnswers: UserAnswers, departureId: String, mode: Mode): Option[Call] =
     (userAnswers.get(IdentificationPage) orElse userAnswers.get(InferredIdentificationPage)).map(_.code).flatMap {
       case CustomsOfficeIdentifier       => Some(controllers.locationOfGoods.routes.CustomsOfficeIdentifierController.onPageLoad(departureId, mode))
       case EoriNumberIdentifier          => Some(controllers.locationOfGoods.routes.EoriController.onPageLoad(departureId, mode))
@@ -81,7 +78,7 @@ class LocationOfGoodsNavigator @Inject() () extends Navigator {
       case _                             => None
     }
 
-  def locationOfGoodsNavigation(ua: UserAnswers, departureId: String, mode: Mode): Option[Call] =
+  private def locationOfGoodsNavigation(ua: UserAnswers, departureId: String, mode: Mode): Option[Call] =
     ua.departureData.Consignment.LocationOfGoods match {
       case None    => Some(controllers.locationOfGoods.routes.LocationTypeController.onPageLoad(departureId, mode))
       case Some(_) => placeOfLoadingExistsRedirect(ua, departureId, mode)
@@ -117,15 +114,5 @@ class LocationOfGoodsNavigator @Inject() () extends Navigator {
     userAnswers.departureData.Consignment.PlaceOfLoading match {
       case Some(_) => locationPageNavigation(departureId, mode, userAnswers)
       case None    => AddUnLocodePage.route(userAnswers, departureId, mode)
-    }
-}
-
-object LocationOfGoodsNavigator {
-
-  def limitDatePageNavigator(departureId: String, mode: Mode, ua: UserAnswers): Option[Call] =
-    if (isContainerIndicatorMissing(ua, mode)) {
-      ContainerIndicatorPage.route(ua, departureId, mode)
-    } else {
-      containerIndicatorPageNavigation(departureId, mode, ua)
     }
 }
